tests/wpt/web-platform-tests/html/canvas/offscreen/line-styles/2d.line.cap.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.line.cap.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.line.cap.valid</h1>
<p class="desc">Setting lineCap to valid values works</p>


<script>
var t = async_test("Setting lineCap to valid values works");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.lineCap = 'butt'
_assertSame(ctx.lineCap, 'butt', "ctx.lineCap", "'butt'");
ctx.lineCap = 'round';
_assertSame(ctx.lineCap, 'round', "ctx.lineCap", "'round'");
ctx.lineCap = 'square';
_assertSame(ctx.lineCap, 'square', "ctx.lineCap", "'square'");
t.done();

});
</script>
